requirements.txt
flask>=3.0
pyswisseph>=2.10.3
supabase>=2.4.0
python-dotenv>=1.0.0
